// app.client/src/api/models/NoteListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type NoteListVM = {
    noteID?: string | null;
    noteText?: string | null;
    isActive?: boolean;
    category?: string | null;
    categoryID?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};